refactor(response): replace ResponseBuilder with ApiResponseBuilder for consistent API responses refactor(user): remove unused Slf4j annotation and improve method structure refactor(exception): update ApiExceptionHandler package and remove deprecated RestExceptionHandler and GlobalExceptionHandler
