--- src/main/java/info/mackiewicz/bankapp/shared/exception/handlers/RestExceptionHandler.java
@@ -1,40 +1,0 @@
-package info.mackiewicz.bankapp.shared.exception.handlers;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import info.mackiewicz.bankapp.account.exception.AccountNotFoundByIdException;
-import info.mackiewicz.bankapp.account.exception.OwnerAccountsNotFoundException;
-import info.mackiewicz.bankapp.presentation.exception.InvalidUserException;
-import info.mackiewicz.bankapp.user.exception.UserValidationException;
-
-@RestControllerAdvice
-public class RestExceptionHandler {
-
-    @ExceptionHandler(InvalidUserException.class)
-    public ResponseEntity<String> handleInvalidUserException(InvalidUserException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-    
-    @ExceptionHandler(UserValidationException.class)
-    public ResponseEntity<String> handleUserValidationException(UserValidationException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-    
-    @ExceptionHandler(AccountNotFoundByIdException.class)
-    public ResponseEntity<String> handleAccountNotFoundByIdException(AccountNotFoundByIdException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-    
-    @ExceptionHandler(OwnerAccountsNotFoundException.class)
-    public ResponseEntity<String> handleOwnerAccountsNotFoundException(OwnerAccountsNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> handleAllExceptions(Exception e) {
-        return ResponseEntity.internalServerError().body("An unexpected error occurred");
-    }
-}
