+++ src/main/java/info/mackiewicz/bankapp/user/service/UserService.java
@@ -23,7 +23,6 @@ import lombok.extern.slf4j.Slf4j;
  * @see UserOperationsService
  */
 @RequiredArgsConstructor
-@Slf4j
 @Service
 public class UserService {
